CardsGrid logic moved from page to component
